+++ CHANGELOG.md
@@ -1,1 +1,14 @@
 # Changelog
+
+## [Unreleased]
+
+## [0.1.0] - 2021-12-01
+
+### Added
+
+-   Draft of `dp init`, `dp create`, `dp template new`, `dp template list` and `dp dbt`
+-   Draft of `dp compile` and `dp deploy`
+
+[Unreleased]: https://github.com/getindata/data-pipelines-cli/compare/0.1.0...HEAD
+
+[0.1.0]: https://github.com/getindata/data-pipelines-cli/compare/5df87160e660d372a52a3665fe79be2029089613...0.1.0